server/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
async-timeout==5.0.1
billiard==4.2.2
celery==5.5.3
certifi==2025.10.5
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
coverage==7.12.0
distro==1.9.0
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
exceptiongroup==1.3.0
factory_boy==3.3.3
Faker==38.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jiter==0.11.1
kombu==5.5.4
mysqlclient==2.2.7
openai==2.7.1
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.52
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.1
pytest-cov==7.0.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.0.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.3.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
